requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonify==0.5
MarkupSafe==1.1.1
Pillow==7.2.0
requests==2.25.1
SQLAlchemy==1.3.22
SQLAlchemy-serializer==1.3.4.4
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
